Move entities into separate library and upgrade to EF 7.0
