+++ server/src/Korga.Core/ChurchTools/Entities/GroupRole.cs
@@ -1,0 +1,18 @@
+﻿namespace Korga.ChurchTools.Entities;
+
+public class GroupRole
+{
+    public GroupRole(int id, int groupTypeId, string name)
+    {
+        Id = id;
+        GroupTypeId = groupTypeId;
+        Name = name;
+    }
+
+    public int Id { get; set; }
+
+    public GroupType? GroupType { get; set; }
+    public int GroupTypeId { get; set; }
+
+    public string Name { get; set; }
+}
