+++ server/src/Korga.Core/ChurchTools/Entities/GroupType.cs
@@ -1,0 +1,13 @@
+﻿namespace Korga.ChurchTools.Entities;
+
+public class GroupType
+{
+    public GroupType(int id, string name)
+    {
+        Id = id;
+        Name = name;
+    }
+
+    public int Id { get; set; }
+    public string Name { get; set; }
+}
