+++ server/src/Korga.Core/EmailRelay/Entities/PersonFilter.cs
@@ -1,0 +1,6 @@
+﻿namespace Korga.EmailRelay.Entities;
+
+public abstract class PersonFilter
+{
+    public long Id { get; set; }
+}
