--- server/src/Korga.Server/Database/EmailRelay/GroupFilter.cs
@@ -1,12 +1,0 @@
-﻿using Korga.Server.Database.ChurchTools;
-
-namespace Korga.Server.Database.EmailRelay;
-
-public class GroupFilter : PersonFilter
-{
-	public Group? Group { get; set; }
-	public int GroupId { get; set; }
-
-	public GroupRole? GroupRole { get; set; }
-	public int? GroupRoleId { get; set; }
-}
