--- server/src/Korga.Server/Database/EmailRelay/PersonFilter.cs
@@ -1,6 +1,0 @@
-﻿namespace Korga.Server.Database.EmailRelay;
-
-public abstract class PersonFilter
-{
-	public long Id { get; set; }
-}
